requirements.txt
# Project Chimera - Enterprise Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
playwright==1.40.0
beautifulsoup4==4.12.2
pyautogen==0.9.0
chromadb==0.4.18
lancedb==0.3.4
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
aiosqlite==0.19.0
requests==2.31.0
pandas==2.1.3
plotly==5.17.0
websocket-client==1.6.4
rel==0.4.9
openai==1.3.7
anthropic==0.7.8
docker==6.1.3

# Enterprise Security Dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1

# Production Dependencies
redis==5.0.1
celery==5.3.4
loguru==0.7.2

# Production Deployment
gunicorn==21.2.0
psycopg2-binary==2.9.9

# Monitoring & Metrics
prometheus-client==0.19.0
psutil==5.9.6

# Multi-tenant Architecture
asyncpg==0.29.0
